Re-enable vector rotates
Closes #63 .
